+++ src/admin/components/Routes/Project/index.tsx
@@ -3,8 +3,10 @@ import { useTranslation } from 'react-i18next';
 import lazy from '../../../utilities/lazy.js';
 import { Loader } from '../../elements/Loader/index.js';
 import { MainHeader } from '../../elements/MainHeader/index.js';
-import { ActivityBarLayout } from '../../layouts/ActivityBarLayout/index.js';
 
+const ActivityBarLayout = Loader(
+  lazy(() => import('../../layouts/ActivityBarLayout/index.js'), 'ActivityBarLayout')
+);
 const CreateProjectPage = Loader(
   lazy(() => import('../../../pages/Project/Create/index.js'), 'CreateProjectPage')
 );
