--- src/admin/components/Routes/Review/index.tsx
@@ -1,36 +1,0 @@
-import React from 'react';
-import { useTranslation } from 'react-i18next';
-import lazy from '../../../utilities/lazy.js';
-import { Loader } from '../../elements/Loader/index.js';
-import { MainHeader } from '../../elements/MainHeader/index.js';
-import { NavContentLayout } from '../../layouts/NavContentLayout/index.js';
-import { useAuth } from '../../utilities/Auth/index.js';
-
-const ReviewDetailPage = Loader(
-  lazy(() => import('../../../pages/Review/Detail/index.js'), 'ReviewDetailPage')
-);
-const NotFound = Loader(lazy(() => import('../../../pages/NotFound/index.js'), 'NotFound'));
-
-export const ReviewRoutes = () => {
-  const { hasPermission } = useAuth();
-  const { t } = useTranslation();
-
-  const children = [{ path: '*', element: <NotFound /> }];
-
-  if (hasPermission('readOwnReview') || hasPermission('readAllReview')) {
-    children.push({
-      path: ':id',
-      element: (
-        <MainHeader label={t('review')}>
-          <ReviewDetailPage />
-        </MainHeader>
-      ),
-    });
-  }
-
-  return {
-    path: '/admin/reviews',
-    element: <NavContentLayout variable="tenant" />,
-    children,
-  };
-};
